Board classes have been moved to a more specific packagesx
