--- src/com/ucm/dasi/catan/board/IBoardConnection.java
@@ -1,5 +1,0 @@
-package com.ucm.dasi.catan.board;
-
-public interface IBoardConnection extends IBoardElement {
-	ConnectionType getType();
-}
