--- src/com/ucm/dasi/catan/board/IBoardStructure.java
@@ -1,5 +1,0 @@
-package com.ucm.dasi.catan.board;
-
-public interface IBoardStructure extends IBoardElement {
-	StructureType getType();
-}
