+++ src/com/ucm/dasi/catan/board/ICatanBoard.java
@@ -1,5 +1,9 @@
 package com.ucm.dasi.catan.board;
 
+import com.ucm.dasi.catan.board.connection.IBoardConnection;
+import com.ucm.dasi.catan.board.structure.IBoardStructure;
+import com.ucm.dasi.catan.board.terrain.IBoardTerrain;
+
 public interface ICatanBoard {
 	IBoardElement get(int x, int y);
