+++ src/com/ucm/dasi/catan/board/connection/ConnectionType.java
@@ -1,4 +1,4 @@
-package com.ucm.dasi.catan.board;
+package com.ucm.dasi.catan.board.connection;
 
 public enum ConnectionType {
 	Road,
