+++ src/com/ucm/dasi/catan/board/connection/IBoardConnection.java
@@ -1,0 +1,7 @@
+package com.ucm.dasi.catan.board.connection;
+
+import com.ucm.dasi.catan.board.IBoardElement;
+
+public interface IBoardConnection extends IBoardElement {
+	ConnectionType getType();
+}
