+++ src/com/ucm/dasi/catan/board/structure/IBoardStructure.java
@@ -1,0 +1,7 @@
+package com.ucm.dasi.catan.board.structure;
+
+import com.ucm.dasi.catan.board.IBoardElement;
+
+public interface IBoardStructure extends IBoardElement {
+	StructureType getType();
+}
